requirements.txt
Django==1.9.1
django-storages==1.1.8
boto==2.38.0
celery==3.1.19
waitress==0.8.9
python-binary-memcached
django-bmemcached
psycopg2==2.6.1
dj-static==0.0.6
pillow
dj-database-url==0.3.0
requests==2.9.0
django-pin-passcode==0.1.5
python-dateutil==2.4.2
